[OPTIMIZE] Add optimizations from the code review : singleton, LocalDate, return types

+++ src/main/java/com/excilys/cdb/model/Company.java
@@ -11,7 +11,7 @@ public class Company {
 	/**
 	 * The identifier of the company
 	 */
-	int id;
+	long id;
 	
 	/**
 	 * The name of the company
@@ -21,20 +21,20 @@ public class Company {
 	public Company() {
 	}
 
-	public Company(int id) {
+	public Company(long id) {
 		this.id = id;
 	}
 	
-	public Company(int id, String name) {
+	public Company(long id, String name) {
 		this.id = id;
 		this.name = name;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 
@@ -55,7 +55,7 @@ public class Company {
 	public int hashCode() {
 		final int prime = 31;
 		int result = 1;
-		result = prime * result + id;
+		result = prime * result + (int)id;
 		result = prime * result + ((name == null) ? 0 : name.hashCode());
 		return result;
 	}
